requirements.txt
streamlit>=1.40.0
python-dotenv>=1.0.0
azure-ai-documentintelligence>=1.0.0
azure-search-documents>=11.6.0
openai>=1.60.0
azure-core>=1.32.0
azure-cognitiveservices-speech>=1.41.0
azure-ai-translation-text>=1.0.0
